+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.SPOSiteScript.Tests.ps1
@@ -55,6 +55,9 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 }
             }
 
+            Mock -CommandName Remove-PnPSiteScript -MockWith {
+            }
+
             Mock -CommandName Start-Sleep -MockWith {
             }
 
@@ -81,12 +84,22 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
                 Mock -CommandName Add-PnPSiteScript -MockWith {
                     return @{
-                        Name = $null
+                        Id = '12345-67890-abcde-f0123'
                     }
                 }
 
-                Mock -CommandName Get-PnPSiteScript -MockWith {
+                # calls to Get-PnPSiteScript without proper Identity returns nothing
+                Mock -CommandName Get-PnPSiteScript -ParameterFilter {Identity -ne '12345-67890-abcde-f0123'} -MockWith {
                     return $null
+                }
+                # after Add-PnPSiteScript has been called, Get-PnPSiteScript should return the created site-script
+                Mock -CommandName Get-PnPSiteScript -ParameterFilter {Identity -eq '12345-67890-abcde-f0123'} -MockWith {
+                    return @{
+                        Id          = '12345-67890-abcde-f0123'
+                        Title       = 'Title One'
+                        Content     = $script
+                        Description = "This is the description for the Site Script: 'Test Title'"
+                    }
                 }
             }
 
@@ -100,6 +113,37 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Creates the site script in the Set method' {
                 Set-TargetResource @testParams
+                Should -Invoke -CommandName Add-PnPSiteScript -Exactly -Times 1
+            }
+        }
+
+        Context -Name 'The site script already exist and is in the desired state' -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    Identity    = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
+                    Title       = 'Title One'
+                    Content     = $script
+                    Description = "This is the description for the Site Script: 'Test Title'"
+                    Credential  = $Credential
+                    Ensure      = 'Present'
+                }
+
+                Mock -CommandName Get-PnPSiteScript -MockWith {
+                    return @{
+                        Id          = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
+                        Title       = 'Title One'
+                        Content     = $script
+                        Description = "This is the description for the Site Script: 'Test Title'"
+                    }
+                }
+            }
+
+            It 'Should return present from the Get method' {
+                (Get-TargetResource @testParams).Ensure | Should -Be 'Present'
+            }
+
+            It 'Should return true from the Test method' {
+                Test-TargetResource @testParams | Should -Be $true
             }
         }
 
@@ -116,14 +160,11 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
                 Mock -CommandName Get-PnPSiteScript -MockWith {
                     return @{
-                        Identity    = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
-                        Title = 'Title One'
+                        Id          = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
+                        Title       = 'Title One'
+                        Description = "wrong description"
                     }
                 }
-            }
-
-            It 'Should update the site script from the Set method' {
-                Set-TargetResource @testParams
             }
 
             It 'Should return present from the Get method' {
@@ -132,6 +173,11 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should return false from the Test method' {
                 Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It 'Should update the site script from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Set-PnPSiteScript -Exactly -Times 1
             }
         }
 
@@ -149,20 +195,16 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     return @(
                         @{
                             Identity    = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
-                            Title = 'Title One'
+                            Title       = 'Title One'
                             Description = "This is the first sitescript with title 'Title One'"
                         },
                         @{
                             Identity    = '01234567-890a-bcde-f094-e8df5f9aa202'
-                            Title = 'Title One'
+                            Title       = 'Title One'
                             Description = "This is another sitescript with the same title"
                         }
                     )
                 }
-            }
-
-            It 'Should update the site script from the Set method' {
-                Set-TargetResource @testParams
             }
 
             It 'Should return present from the Get method' {
@@ -171,6 +213,11 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should return false from the Test method' {
                 Test-TargetResource @testParams | Should -Be $false
+            }
+
+            It 'Should update the site script from the Set method' {
+                Set-TargetResource @testParams
+                Should -Invoke -CommandName Set-PnPSiteScript -Exactly -Times 1
             }
         }
 
@@ -188,7 +235,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                 Mock -CommandName Get-PnPSiteScript -MockWith {
                     return @{
                         Id          = '6c42cc50-7f90-45c2-9094-e8df5f9aa202'
-                        Title       = 'Title Two'
+                        Title       = 'Title One'
                         Content     = $script
                         Description = "This is the description for the Site Script: 'Test Title'"
                     }
@@ -201,6 +248,7 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
 
             It 'Should remove the site script successfully' {
                 Set-TargetResource @testParams
+                Should -Invoke -CommandName Remove-PnPSiteScript -Exactly -Times 1
             }
         }