+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCPolicyConfig/MSFT_SCPolicyConfig.psm1
@@ -63,6 +63,10 @@ function Get-TargetResource
 
         [Parameter()]
         [System.Boolean]
+        $FileCopiedToCloudFullUrlEnabled,
+
+        [Parameter()]
+        [System.Boolean]
         $IncludePredefinedUnallowedBluetoothApps,
 
         [Parameter()]
@@ -177,19 +181,28 @@ function Get-TargetResource
         $DlpNetworkShareGroupsObject = ConvertFrom-Json $instance.DlpNetworkShareGroups
 
         # AdvancedClassificationEnabled
-        $AdvancedClassificationEnabledValue = [Boolean]::Parse(($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'AdvancedClassificationEnabled' }).Value)
+        $AdvancedClassificationEnabledValue = $false # default value
+        $valueToParse =($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'AdvancedClassificationEnabled' }).Value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
+        {
+            $AdvancedClassificationEnabledValue = [Boolean]::Parse($valueToParse)
+        }
 
         # BandwidthLimitEnabled
-        $toBeParsed = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'BandwidthLimitEnabled' }).Value
-        $parsedValue = $null
-        if ($null -ne $toBeParsed)
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'BandwidthLimitEnabled' }).Value
+        $BandwidthLimitEnabledValue = $true #default value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
         {
-            $parsedValue = [Boolean]::Parse($toBeParsed)
+            $BandwidthLimitEnabledValue = [Boolean]::Parse($valueToParse)
         }
-        $BandwidthLimitEnabledValue = $parsedValue
 
         # DailyBandwidthLimitInMB
-        $DailyBandwidthLimitInMBValue = [UInt32]($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'DailyBandwidthLimitInMB' }).Value
+        $DailyBandwidthLimitInMBValue = 1000 # default value
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'DailyBandwidthLimitInMB' }).Value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
+        {
+            $DailyBandwidthLimitInMBValue = [UInt32]$valueToParse
+        }
 
         # PathExclusion
         $PathExclusionValue = [Array]($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'PathExclusion' }).Value
@@ -198,7 +211,12 @@ function Get-TargetResource
         $MacPathExclusionValue = [Array]($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'MacPathExclusion' }).Value
 
         # MacDefaultPathExclusionsEnabled
-        $MacDefaultPathExclusionsEnabledValue = [Boolean]::Parse(($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'MacDefaultPathExclusionsEnabled' }).Value)
+        $MacDefaultPathExclusionsEnabledValue = $true # default value
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'MacDefaultPathExclusionsEnabled' }).Value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
+        {
+            $MacDefaultPathExclusionsEnabledValue = [Boolean]::Parse($valueToParse)
+        }
 
         #EvidenceStoreSettings
         $entry = $EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'EvidenceStoreSettings' }
@@ -214,7 +232,12 @@ function Get-TargetResource
         }
 
         # NetworkPathEnforcementEnabled
-        $NetworkPathEnforcementEnabledValue = [Boolean]::Parse(($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'NetworkPathEnforcementEnabled' }).Value)
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'NetworkPathEnforcementEnabled' }).Value
+        $NetworkPathEnforcementEnabledValue = $false # default value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
+        {
+            $NetworkPathEnforcementEnabledValue = [Boolean]::Parse($valueToParse)
+        }
 
         # NetworkPathExclusion
         $NetworkPathExclusionValue = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'NetworkPathExclusion' }).Value
@@ -267,13 +290,12 @@ function Get-TargetResource
         }
 
         # IncludePredefinedUnallowedBluetoothApps
-        $toBeParsed = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'IncludePredefinedUnallowedBluetoothApps' }).Value
-        $parsedValue = $null
-        if ($null -ne $toBeParsed)
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'IncludePredefinedUnallowedBluetoothApps' }).Value
+        $IncludePredefinedUnallowedBluetoothAppsValue = $true # default value
+        if (-not [System.String]::IsNullOrEMpty($valueToParse))
         {
-            $parsedValue = [Boolean]::Parse($toBeParsed)
+            $IncludePredefinedUnallowedBluetoothAppsValue = [Boolean]::Parse($valueToParse)
         }
-        $IncludePredefinedUnallowedBluetoothAppsValue = $parsedValue
 
         # UnallowedBluetoothApp
         $entries = [Array]($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'UnallowedBluetoothApp' })
@@ -352,10 +374,20 @@ function Get-TargetResource
             }
 
             # serverDlpEnabled
-            $serverDlpEnabledValue = [Boolean]::Parse(($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'serverDlpEnabled' }).Value)
+            $serverDlpEnabledValue = $false #default value
+            $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'serverDlpEnabled' }).Value
+            if (-not [System.String]::IsNullOrEmpty($valueToParse))
+            {
+                $serverDlpEnabledValue = [Boolean]::Parse($valueToParse)
+            }
 
             # AuditFileActivity
-            $AuditFileActivityValue = [Boolean]::Parse(($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'AuditFileActivity' }).Value)
+            $AuditFileActivityValue = $false # default value
+            $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'AuditFileActivity' }).Value
+            if (-not [System.String]::IsNullOrEmpty($valueToParse))
+            {
+                $AuditFileActivityValue = [Boolean]::Parse($valueToParse)
+            }
 
             # VPNSettings
             $entity = $EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'VPNSettings' }
@@ -454,10 +486,26 @@ function Get-TargetResource
             }
         }
 
+        #EnableLabelCoauthValue
+        $EnableLabelCoauthValue = $false # default value
+        if (-not [System.String]::IsNullOrEmpty($instance.EnableLabelCoauth))
+        {
+            $EnableLabelCoauthValue = $instance.EnableLabelCoauth
+        }
+
+        #FileCopiedToCloudFullUrlEnabledValue
+        $FileCopiedToCloudFullUrlEnabledValue = $false
+        $valueToParse = ($EndpointDlpGlobalSettingsValue | Where-Object { $_.Setting -eq 'FileCopiedToCloudFullUrlEnabled' }).Value
+        if (-not [System.String]::IsNullOrEmpty($valueToParse))
+        {
+            $FileCopiedToCloudFullUrlEnabledValue = [Boolean]::Parse($valueToParse)
+        }
+
         $results = @{
             IsSingleInstance                        = 'Yes'
             AdvancedClassificationEnabled           = $AdvancedClassificationEnabledValue
             BandwidthLimitEnabled                   = $BandwidthLimitEnabledValue
+            FileCopiedToCloudFullUrlEnabled         = $FileCopiedToCloudFullUrlEnabledValue
             DailyBandwidthLimitInMB                 = $DailyBandwidthLimitInMBValue
             PathExclusion                           = $PathExclusionValue
             MacPathExclusion                        = $MacPathExclusionValue
@@ -482,7 +530,7 @@ function Get-TargetResource
             DLPRemovableMediaGroups                 = $DLPRemovableMediaGroupsValue
             DLPNetworkShareGroups                   = $DlpNetworkShareGroupsValue
             VPNSettings                             = $VPNSettingsValue
-            EnableLabelCoauth                       = $instance.EnableLabelCoauth
+            EnableLabelCoauth                       = $EnableLabelCoauthValue
             EnableSpoAipMigration                   = $instance.EnableSpoAipMigration
             QuarantineParameters                    = $QuarantineParametersValue
             Credential                              = $Credential
@@ -568,6 +616,10 @@ function Set-TargetResource
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $EvidenceStoreSettings,
+
+        [Parameter()]
+        [System.Boolean]
+        $FileCopiedToCloudFullUrlEnabled,
 
         [Parameter()]
         [System.Boolean]
@@ -1091,6 +1143,10 @@ function Test-TargetResource
         [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $EvidenceStoreSettings,
+
+        [Parameter()]
+        [System.Boolean]
+        $FileCopiedToCloudFullUrlEnabled,
 
         [Parameter()]
         [System.Boolean]